+++ homeassistant/components/tahoma/sensor.py
@@ -5,24 +5,27 @@ import logging
 from homeassistant.const import ATTR_BATTERY_LEVEL, TEMP_CELSIUS, UNIT_PERCENTAGE
 from homeassistant.helpers.entity import Entity
 
-from . import DOMAIN as TAHOMA_DOMAIN, TahomaDevice
+from .const import DOMAIN, TAHOMA_TYPES
+from .tahoma_device import TahomaDevice
 
 _LOGGER = logging.getLogger(__name__)
 
 SCAN_INTERVAL = timedelta(seconds=60)
 
-ATTR_RSSI_LEVEL = "rssi_level"
 
+async def async_setup_entry(hass, entry, async_add_entities):
+    """Set up the Tahoma sensors from a config entry."""
 
-def setup_platform(hass, config, add_entities, discovery_info=None):
-    """Set up Tahoma controller devices."""
-    if discovery_info is None:
-        return
-    controller = hass.data[TAHOMA_DOMAIN]["controller"]
-    devices = []
-    for device in hass.data[TAHOMA_DOMAIN]["devices"]["sensor"]:
-        devices.append(TahomaSensor(device, controller))
-    add_entities(devices, True)
+    data = hass.data[DOMAIN][entry.entry_id]
+
+    entities = []
+    controller = data.get("controller")
+
+    for device in data.get("devices"):
+        if TAHOMA_TYPES[device.uiclass] == "sensor":
+            entities.append(TahomaSensor(device, controller))
+
+    async_add_entities(entities)
 
 
 class TahomaSensor(TahomaDevice, Entity):
@@ -42,20 +45,17 @@ class TahomaSensor(TahomaDevice, Entity):
     @property
     def unit_of_measurement(self):
         """Return the unit of measurement of this entity, if any."""
-        if self.tahoma_device.type == "io:TemperatureIOSystemSensor":
+
+        if self.tahoma_device.uiclass == "TemperatureSensor":
             return TEMP_CELSIUS
-        if self.tahoma_device.type == "io:SomfyContactIOSystemSensor":
-            return None
-        if self.tahoma_device.type == "io:SomfyBasicContactIOSystemSensor":
-            return None
-        if self.tahoma_device.type == "io:LightIOSystemSensor":
-            return "lx"
-        if self.tahoma_device.type == "Humidity Sensor":
+
+        if self.tahoma_device.uiclass == "HumiditySensor":
             return UNIT_PERCENTAGE
-        if self.tahoma_device.type == "rtds:RTDSContactSensor":
-            return None
-        if self.tahoma_device.type == "rtds:RTDSMotionSensor":
-            return None
+
+        if self.tahoma_device.uiclass == "LightSensor":
+            return "lx"
+
+        return None
 
     def update(self):
         """Update the state."""
@@ -88,26 +88,3 @@ class TahomaSensor(TahomaDevice, Entity):
             self._available = True
 
         _LOGGER.debug("Update %s, value: %d", self._name, self.current_value)
-
-    @property
-    def device_state_attributes(self):
-        """Return the device state attributes."""
-        attr = {}
-        super_attr = super().device_state_attributes
-        if super_attr is not None:
-            attr.update(super_attr)
-
-        if "core:RSSILevelState" in self.tahoma_device.active_states:
-            attr[ATTR_RSSI_LEVEL] = self.tahoma_device.active_states[
-                "core:RSSILevelState"
-            ]
-        if "core:SensorDefectState" in self.tahoma_device.active_states:
-            attr[ATTR_BATTERY_LEVEL] = self.tahoma_device.active_states[
-                "core:SensorDefectState"
-            ]
-        return attr
-
-    @property
-    def available(self):
-        """Return True if entity is available."""
-        return self._available
