+++ homeassistant/components/tahoma/strings.json
