+++ homeassistant/components/tahoma/tahoma_device.py
