+++ homeassistant/components/tahoma/translations/en.json
