--- homeassistant/components/tahoma_refactor/manifest.json
@@ -1,8 +1,0 @@
-{
-  "domain": "tahoma_refactor",
-  "name": "Tahoma",
-  "config_flow": true,
-  "documentation": "https://www.home-assistant.io/integrations/tahoma_refactor",
-  "requirements": [],
-  "codeowners": ["@imicknl", "@vlebourl"]
-}
